+++ apps/web/components/Subscription.tsx
@@ -7,7 +7,7 @@ import {
   Code,
   useColorMode,
 } from "@chakra-ui/react";
-import { useSubscription, useWeb3InboxClient } from "@web3inbox/widget-react";
+import { useSubscription, useWeb3InboxClient } from "@web3inbox/react";
 import React from "react";
 
 function Subscription() {
@@ -23,7 +23,7 @@ function Subscription() {
     <AccordionItem border="none">
       <AccordionButton py="4">
         <Heading fontSize="md" as="span" flex="1" textAlign="left">
-          Subscription
+          Subscription (useSubscription)
         </Heading>
         <AccordionIcon />
       </AccordionButton>