+++ packages/core/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @web3inbox/core
 
+## 1.0.2
+
+### Patch Changes
+
+- Fix bug where paging notifications was sometimes a noop
+
+## 1.0.1
+
+### Patch Changes
+
+- Fix bug with duplicate notifications
+
+## 1.0.0
+
+### Major Changes
+
+- Stable release
+
 ## 0.0.25
 
 ### Patch Changes