+++ packages/react/CHANGELOG.md
@@ -1,4 +1,29 @@
-# @web3inbox/widget-react
+# @web3inbox/react
+
+## 1.0.2
+
+### Patch Changes
+
+- Fix bug where paging notifications was sometimes a noop
+- Updated dependencies
+  - @web3inbox/core@1.0.2
+
+## 1.0.1
+
+### Patch Changes
+
+- Update README and docs
+
+## 1.0.0
+
+### Major Changes
+
+- Stable release
+
+### Patch Changes
+
+- Updated dependencies
+  - @web3inbox/core@1.0.0
 
 ## 0.6.26